# Conflicts:
#	backend/src/main/resources/application.properties

+++ backend/src/main/java/com/ssafish/common/util/WebSocketSubscriberManager.java
@@ -13,12 +13,12 @@ import java.util.concurrent.ConcurrentHashMap;
 @Component
 public class WebSocketSubscriberManager {
 
-    private final Map<Integer, String> entireSessions = new ConcurrentHashMap<>();
-    private final Map<Integer, Map<Integer, String>> rooms = new ConcurrentHashMap<>();
-    private final Map<String, Integer> roomTable = new ConcurrentHashMap<>();
-    private final Map<String, Integer> userTable = new ConcurrentHashMap<>();
+    private final Map<Long, String> entireSessions = new ConcurrentHashMap<>();
+    private final Map<Long, Map<Long, String>> rooms = new ConcurrentHashMap<>();
+    private final Map<String, Long> roomTable = new ConcurrentHashMap<>();
+    private final Map<String, Long> userTable = new ConcurrentHashMap<>();
 
-    public void addSubscriber(Integer roomId, Integer userId, String sessionId) {
+    public void addSubscriber(Long roomId, Long userId, String sessionId) {
         entireSessions.put(userId, sessionId);
         if (!rooms.containsKey(roomId)) {
             rooms.put(roomId, new ConcurrentHashMap<>());
@@ -30,8 +30,8 @@ public class WebSocketSubscriberManager {
 
     public void removeSubscriber(String sessionId) {
         if (roomTable.containsKey(sessionId) && userTable.containsKey(sessionId)) {
-            Integer roomId = roomTable.get(sessionId);
-            Integer userId = userTable.get(sessionId);
+            Long roomId = roomTable.get(sessionId);
+            Long userId = userTable.get(sessionId);
             entireSessions.remove(userId);
             rooms.get(roomId).remove(userId);
             roomTable.remove(sessionId);
@@ -39,18 +39,18 @@ public class WebSocketSubscriberManager {
         }
     }
 
-    public Map<Integer, String> getRoomById(Integer roomId) {
+    public Map<Long, String> getRoomById(Long roomId) {
         return rooms.get(roomId);
     }
     public boolean isExist(String sessionId) {
         return roomTable.containsKey(sessionId);
     }
 
-    public Integer getRoomIdBySessionId(String sessionId) {
+    public Long getRoomIdBySessionId(String sessionId) {
         return roomTable.get(sessionId);
     }
 
-    public Integer getUserIdBySessionId(String sessionId) {
+    public Long getUserIdBySessionId(String sessionId) {
         return userTable.get(sessionId);
     }
 }
